+++ docs/CHANGELOG.rst
@@ -5,6 +5,14 @@ Change Log
 All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
+==========
+Unreleased
+==========
+
+Added
+-----
+- Add getSpeciesFromFeatures utility function
+
 ==================
 2.0.2 - 2017-11-03
 ==================